+++ src/xyz/oihana/schema/constants/traits/products/PhysicalQuantity.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\products ;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\products\PhysicalQuantity} class.
+ *
+ * @author  Marc Alcaraz (eKameleon)
+ * @package xyz\oihana\schema\constants\traits\products
+ * @since   1.4.0
+ */
+trait PhysicalQuantity
+{
+    public const string VOLUME = 'volume' ;
+    public const string WEIGHT = 'weight' ;
+}